huggingface_space/requirements.txt
gradio==5.12.0
spaces
numpy
Pillow
torch
torchvision
diffusers
transformers
accelerate
safetensors
peft
